Provide redirect API publicly
